creacion de funciones utils para el manejo de cloudinary
